fix(project): fix font family
add font family to the layout

fixes #40
